+++ pkg/configs/api/api.go
@@ -3,6 +3,7 @@ package api
 import (
 	"database/sql"
 	"encoding/json"
+	"flag"
 	"fmt"
 	"html/template"
 	"io/ioutil"
@@ -19,15 +20,47 @@ import (
 	"github.com/cortexproject/cortex/pkg/util"
 )
 
+type Config struct {
+	Notifications NotificationsConfig `yaml:"notifications"`
+}
+
+type NotificationsConfig struct {
+	DisableEmail     bool `yaml:"allow_email"`
+	DisablePagerDuty bool `yaml:"allow_pagerduty"`
+	DisablePushover  bool `yaml:"allow_pushover"`
+	DisableSlack     bool `yaml:"allow_slack"`
+	DisableOpsGenie  bool `yaml:"allow_opsgenie"`
+	DisableWebHook   bool `yaml:"allow_webhook"`
+	DisableVictorOps bool `yaml:"allow_victorops"`
+	DisableWeChat    bool `yaml:"allow_wechat"`
+	// Hipchat has reached end of life and is no longer available
+}
+
+// RegisterFlagsWithPrefix adds the flags required to config this to the given FlagSet
+func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
+	f.BoolVar(&cfg.Notifications.DisableEmail, "configs-api.notifications.disable-email", false, "Disable Email notifications for Alertmanager.")
+	f.BoolVar(&cfg.Notifications.DisablePagerDuty, "configs-api.notifications.disable-pagerduty", false, "Disable PagerDuty notifications for Alertmanager.")
+	f.BoolVar(&cfg.Notifications.DisablePushover, "configs-api.notifications.disable-pushover", false, "Disable Pushover notifications for Alertmanager.")
+	f.BoolVar(&cfg.Notifications.DisableSlack, "configs-api.notifications.disable-slack", false, "Disable Slack notifications for Alertmanager.")
+	f.BoolVar(&cfg.Notifications.DisableOpsGenie, "configs-api.notifications.disable-opsgenie", false, "Disable OpsGenie notifications for Alertmanager.")
+	f.BoolVar(&cfg.Notifications.DisableWebHook, "configs-api.notifications.disable-webhook", false, "Disable WebHook notifications for Alertmanager.")
+	f.BoolVar(&cfg.Notifications.DisableVictorOps, "configs-api.notifications.disable-victorops", false, "Disable VictorOps notifications for Alertmanager.")
+	f.BoolVar(&cfg.Notifications.DisableWeChat, "configs-api.notifications.disable-wechat", false, "Disable WeChat notifications for Alertmanager.")
+}
+
 // API implements the configs api.
 type API struct {
-	db db.DB
 	http.Handler
+	db  db.DB
+	cfg Config
 }
 
 // New creates a new API
-func New(database db.DB) *API {
-	a := &API{db: database}
+func New(database db.DB, cfg Config) *API {
+	a := &API{
+		db:  database,
+		cfg: cfg,
+	}
 	r := mux.NewRouter()
 	a.RegisterRoutes(r)
 	a.Handler = r
@@ -117,7 +150,7 @@ func (a *API) setConfig(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
-	if err := validateAlertmanagerConfig(cfg.AlertmanagerConfig); err != nil && cfg.AlertmanagerConfig != "" {
+	if err := validateAlertmanagerConfig(cfg.AlertmanagerConfig, a.cfg.Notifications); err != nil && cfg.AlertmanagerConfig != "" {
 		level.Error(logger).Log("msg", "invalid Alertmanager config", "err", err)
 		http.Error(w, fmt.Sprintf("Invalid Alertmanager config: %v", err), http.StatusBadRequest)
 		return
@@ -150,7 +183,7 @@ func (a *API) validateAlertmanagerConfig(w http.ResponseWriter, r *http.Request)
 		return
 	}
 
-	if err = validateAlertmanagerConfig(string(cfg)); err != nil {
+	if err = validateAlertmanagerConfig(string(cfg), a.cfg.Notifications); err != nil {
 		w.WriteHeader(http.StatusBadRequest)
 		util.WriteJSONResponse(w, map[string]string{
 			"status": "error",
@@ -164,15 +197,36 @@ func (a *API) validateAlertmanagerConfig(w http.ResponseWriter, r *http.Request)
 	})
 }
 
-func validateAlertmanagerConfig(cfg string) error {
+func validateAlertmanagerConfig(cfg string, noCfg NotificationsConfig) error {
 	amCfg, err := amconfig.Load(cfg)
 	if err != nil {
 		return err
 	}
 
 	for _, recv := range amCfg.Receivers {
-		if len(recv.EmailConfigs) != 0 {
-			return fmt.Errorf("email notifications are not supported in Cortex yet")
+		if noCfg.DisableEmail && len(recv.EmailConfigs) > 0 {
+			return fmt.Errorf("email notifications are disabled in Cortex yet")
+		}
+		if noCfg.DisablePagerDuty && len(recv.PagerdutyConfigs) > 0 {
+			return fmt.Errorf("pager-duty notifications are disabled in Cortex yet")
+		}
+		if noCfg.DisablePushover && len(recv.PushoverConfigs) > 0 {
+			return fmt.Errorf("pushover notifications are disabled in Cortex yet")
+		}
+		if noCfg.DisableSlack && len(recv.SlackConfigs) > 0 {
+			return fmt.Errorf("slack notifications are disabled in Cortex yet")
+		}
+		if noCfg.DisableOpsGenie && len(recv.OpsGenieConfigs) > 0 {
+			return fmt.Errorf("ops-genie notifications are disabled in Cortex yet")
+		}
+		if noCfg.DisableWebHook && len(recv.WebhookConfigs) > 0 {
+			return fmt.Errorf("web-hook notifications are disabled in Cortex yet")
+		}
+		if noCfg.DisableVictorOps && len(recv.VictorOpsConfigs) > 0 {
+			return fmt.Errorf("victor-ops notifications are disabled in Cortex yet")
+		}
+		if noCfg.DisableWeChat && len(recv.WechatConfigs) > 0 {
+			return fmt.Errorf("we-chat notifications are disabled in Cortex yet")
 		}
 	}